refactor: split the fuzzer tests into a separate file
